tsmp::reflect specialisation via forward definition matching
This commit introduces forward declarations for reflected types and uses these forward
definitions to do the template specialisation of the tsmp::reflect and tsmp::proxy traits.
This is an improvement over the duck type matching that has been used before because it can
make the trait specialisations match excatly the type that is requested. However, it is not
yet possible to forward declare every type. Enumerations, Nested types and types in anonymous
scopes (e.g. functions) are still being duck type matched.